requirements/production.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile requirements/production.in
#
aiohttp==3.8.3
    # via
    #   -r requirements/production.in
    #   aiohttp-cors
    #   faust-streaming
aiohttp-cors==0.7.0
    # via faust-streaming
aiokafka==0.7.2
    # via faust-streaming
aiosignal==1.2.0
    # via aiohttp
anyio==3.6.1
    # via starlette
async-timeout==4.0.2
    # via
    #   aiohttp
    #   redis
attrs==22.1.0
    # via aiohttp
certifi==2022.9.24
    # via requests
charset-normalizer==2.1.1
    # via
    #   aiohttp
    #   requests
click==7.1.2
    # via
    #   faust-streaming
    #   uvicorn
colorlog==6.7.0
    # via mode-streaming
croniter==1.3.7
    # via
    #   faust-streaming
    #   mode-streaming
cython==0.29.32
    # via vedis
deprecated==1.2.13
    # via redis
dnspython==2.2.1
    # via email-validator
email-validator==1.3.0
    # via pydantic
fastapi==0.85.0
    # via -r requirements/production.in
faust-streaming==0.8.11
    # via -r requirements/production.in
frozenlist==1.3.1
    # via
    #   aiohttp
    #   aiosignal
h11==0.14.0
    # via uvicorn
hvac==0.11.2
    # via sitri
idna==3.4
    # via
    #   anyio
    #   email-validator
    #   requests
    #   yarl
intervaltree==3.1.0
    # via faust-streaming
kafka-python==2.0.2
    # via aiokafka
loguru==0.6.0
    # via -r requirements/production.in
mode-streaming==0.2.1
    # via faust-streaming
motor==3.0.0
    # via -r requirements/production.in
multidict==6.0.2
    # via
    #   aiohttp
    #   yarl
mypy-extensions==0.4.3
    # via
    #   faust-streaming
    #   mode-streaming
numpy==1.23.3
    # via pandas
opentracing==1.3.0
    # via faust-streaming
packaging==21.3
    # via redis
pandas==1.5.0
    # via -r requirements/production.in
pydantic[dotenv,email]==1.9.1
    # via
    #   fastapi
    #   sitri
pymongo==4.2.0
    # via motor
pyparsing==3.0.9
    # via packaging
python-dateutil==2.8.2
    # via
    #   croniter
    #   pandas
python-dotenv==0.21.0
    # via pydantic
pytz==2022.4
    # via pandas
pyyaml==6.0
    # via sitri
redis==4.3.4
    # via sitri
requests==2.28.1
    # via hvac
sitri[all]==2.1.4
    # via -r requirements/production.in
six==1.16.0
    # via
    #   faust-streaming
    #   hvac
    #   python-dateutil
sniffio==1.3.0
    # via anyio
sortedcontainers==2.4.0
    # via intervaltree
starlette==0.20.4
    # via fastapi
stringcase==1.2.0
    # via -r requirements/production.in
structlog==22.1.0
    # via sitri
terminaltables==3.1.10
    # via faust-streaming
toml==0.10.2
    # via sitri
typing-extensions==4.3.0
    # via pydantic
ujson==5.4.0
    # via sitri
urllib3==1.26.12
    # via requests
uvicorn==0.18.3
    # via -r requirements/production.in
vedis==0.7.1
    # via sitri
venusian==3.0.0
    # via faust-streaming
wrapt==1.14.1
    # via deprecated
yarl==1.8.1
    # via
    #   aiohttp
    #   faust-streaming
